Harden omnibar focus against review hazards
The shortcut path still focuses the native omnibar synchronously, but pending focus now carries select-all intent through lazy registration and window attachment. The SwiftUI update fallback records the same focus intent and replays it on the next main turn, avoiding the AppKit text-input reentrancy risk while preserving immediate Cmd+L handling.

The focus registry and debug latency tracker moved out of BrowserPanelView so the CI line-budget guard stays green without accepting extra debt. Immediate Cmd+L/Cmd+Shift+L coverage now lives in a dedicated UI test file below the tracked threshold.

Constraint: Do not run local tests; verify builds only with scripts/reload.sh --tag issue-3470-browser-addressbar-lag --launch

Rejected: Bump the Swift file-length budget | extraction keeps the oversized files under budget without accepting new debt

Confidence: high

Scope-risk: moderate

Directive: Keep Cmd+L and Cmd+Shift+L native-field focus synchronous outside SwiftUI updateNSView; only the update fallback may defer to avoid IME reentrancy

Tested: ./scripts/reload.sh --tag issue-3470-browser-addressbar-lag --launch

Not-tested: Local XCTest/UI test execution is prohibited; GitHub/CircleCI checks will cover tests
